Add files via upload
The landscape of Airbnb rentals in New York City presents a diverse array of insights when examined through the lens of data visualization. The following report leverages data scraped from Airbnb as of December 5, 2021, provided by Inside Airbnb, to uncover patterns and correlations within the market. Through an analysis of various factors such as average prices, room types, booking status, and cancellations, we aim to unearth strategic insights that can guide investment and operational decisions for stakeholders in the Airbnb ecosystem.